third_party/WebKit/LayoutTests/fast/css/invalidation/independent-inheritance-fast-path-initial-keyword.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="outer">
    <div id="inner" style="visibility: initial;"></div>
</div>
<script>
test(function(t)
{
    outer.offsetTop; // Force recalc.
    outer.style.visibility = "hidden";
    assert_equals(internals.updateStyleAndReturnAffectedElementCount(), 1);
    assert_equals(getComputedStyle(inner).visibility, "visible");

    outer.offsetTop; // Force recalc.
    outer.style.visibility = "visible";
    assert_equals(internals.updateStyleAndReturnAffectedElementCount(), 1);
    assert_equals(getComputedStyle(inner).visibility, "visible");

}, "Changing visibility, an independent inherited property, is not propagated when using the 'initial' keyword.");
</script>
